Divide appium_tests into each module tests(android)
